+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+### 0.0.2 (2022-11-28)
+
+#### Features
+
+- Add reporting categories.
+- Delete script files and put all scripts in the popup.js file.
+
 ### 0.0.1 (2022-11-26)
 
 #### Features